requirements.txt
openai
openai-agents
jupyter
jupyterlab_rise
nest_asyncio
streamlit
python-dotenv
ipywidgets
mcp
